+++ tests/testthat/test-enw_expectation.R
@@ -15,7 +15,9 @@ test_that("enw_expectation produces the expected default model", {
       "expr_beta", "expr_beta_sd", "expr_lelatent_int", "expr_r_int",
       "expl_beta", "expl_beta_sd",
       "expr_arima_pacf", "expr_arima_theta", "expr_arima_sigma",
-      "expl_arima_pacf", "expl_arima_theta", "expl_arima_sigma"
+      "expr_gp_rho", "expr_gp_alpha",
+      "expl_arima_pacf", "expl_arima_theta", "expl_arima_sigma",
+      "expl_gp_rho", "expl_gp_alpha"
     )
   )
 })
@@ -117,8 +119,11 @@ test_that(
         expr_beta = 3L, expr_beta_sd = 1L, expr_lelatent_int = c(2L, 3L),
         expr_r_int = 1L, expl_beta = 6L, expl_beta_sd = 3L,
         expr_arima_pacf = NULL, expr_arima_theta = NULL,
-        expr_arima_sigma = NULL, expl_arima_pacf = NULL,
-        expl_arima_theta = NULL, expl_arima_sigma = NULL
+        expr_arima_sigma = NULL,
+        expr_gp_rho = NULL, expr_gp_alpha = NULL,
+        expl_arima_pacf = NULL,
+        expl_arima_theta = NULL, expl_arima_sigma = NULL,
+        expl_gp_rho = NULL, expl_gp_alpha = NULL
       )
     )
   }